simplified ODEs per code review, and updated docs to reflect the change
